backend/requirements.txt
fastapi
fastapi_login
dotenv
requests
pygithub